Split out some helper functions
