+++ vscode_extension/src/configuration.ts
@@ -1,0 +1,5 @@
+import { workspace } from "vscode";
+
+export function getConfigurationItem(name: string): any {
+    return workspace.getConfiguration().get(`aac.${name}`) ?? null;
+}
